requirements.txt
numpy==2.3.1
scikit-learn==1.7.1
pandas
pyyaml
matplotlib
tqdm
